MISSIONS: Add vision proxy config, port some more navigator files back
