+++ command/sub8_missions/sub8/vision_proxies.yaml
@@ -1,0 +1,4 @@
+proxy1:
+    root: "/vision/serviceTest"
+proxy2:
+    root: "/vision/serviceTesting2"
