+++ utils/sub8_ros_tools/sub8_ros_tools/__init__.py
@@ -7,7 +7,7 @@ import geometry_helpers
 import func_helpers
 
 from image_helpers import *
-from init_helper import *
+from init_helpers import *
 from msg_helpers import *
 from threading_helpers import *
 from geometry_helpers import *
